int a float

+++ js/seccion2.js
@@ -14,7 +14,7 @@ d3.csv("data/Sec2.csv").then(function(data) {
 
   for(var i=0; i < data.length; i++)
   {
-    data[i].puntos = parseInt(data[i].nota_padlet);
+    data[i].puntos = parseFloat(data[i].nota_padlet);
   }
 
   data.sort((a, b) => (a.puntos < b.puntos) ? 1 : -1);
